Add logic for storing array w order of operations
Use array to store a number input of any size, then record the following clicked operation. Add the number, and operation chose after, store to array, and wipe the number so the process is repeated on user input, increasing array in the order of number, then operation until its ready for calculation.